improved some stuff

+++ src/script_interpreter_func_fe3d_quad2d_set.cpp
@@ -270,7 +270,7 @@ const bool ScriptInterpreter::_executeFe3dQuad2dSetter(const string& functionNam
 		{
 			if(_validateFe3dQuad2d(args[0]->getString(), false))
 			{
-				_fe3d->quad2d_setTextureRepeat(args[0]->getString(), args[1]->getInteger());
+				_fe3d->quad2d_setTextureRepeat(args[0]->getString(), static_cast<unsigned int>(args[1]->getInteger()));
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::EMPTY));
 			}
@@ -481,7 +481,7 @@ const bool ScriptInterpreter::_executeFe3dQuad2dSetter(const string& functionNam
 					return true;
 				}
 
-				_fe3d->quad2d_setAnimationRowIndex(args[0]->getString(), args[1]->getString(), args[1]->getInteger());
+				_fe3d->quad2d_setAnimationRowIndex(args[0]->getString(), args[1]->getString(), static_cast<unsigned int>(args[1]->getInteger()));
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::EMPTY));
 			}
@@ -501,7 +501,7 @@ const bool ScriptInterpreter::_executeFe3dQuad2dSetter(const string& functionNam
 					return true;
 				}
 
-				_fe3d->quad2d_setAnimationColumnIndex(args[0]->getString(), args[1]->getString(), args[1]->getInteger());
+				_fe3d->quad2d_setAnimationColumnIndex(args[0]->getString(), args[1]->getString(), static_cast<unsigned int>(args[1]->getInteger()));
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::EMPTY));
 			}
@@ -521,7 +521,7 @@ const bool ScriptInterpreter::_executeFe3dQuad2dSetter(const string& functionNam
 					return true;
 				}
 
-				_fe3d->quad2d_setAnimationIntervalMultiplier(args[0]->getString(), args[1]->getString(), args[1]->getInteger());
+				_fe3d->quad2d_setAnimationIntervalMultiplier(args[0]->getString(), args[1]->getString(), static_cast<unsigned int>(args[1]->getInteger()));
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::EMPTY));
 			}
@@ -541,7 +541,7 @@ const bool ScriptInterpreter::_executeFe3dQuad2dSetter(const string& functionNam
 					return true;
 				}
 
-				_fe3d->quad2d_setAnimationIntervalDivider(args[0]->getString(), args[1]->getString(), args[1]->getInteger());
+				_fe3d->quad2d_setAnimationIntervalDivider(args[0]->getString(), args[1]->getString(), static_cast<unsigned int>(args[1]->getInteger()));
 
 				returnValues.push_back(make_shared<ScriptValue>(SVT::EMPTY));
 			}
